Refactor edit
